+++ packages/pangea-sdk/pangea/asyncio/services/ai_guard.py
@@ -58,7 +58,6 @@ class AIGuardAsync(ServiceBaseAsync):
         *,
         recipe: str | None = None,
         debug: bool | None = None,
-        llm_info: str | None = None,
         log_fields: LogFields | None = None,
     ) -> PangeaResponse[TextGuardResult[None]]:
         """
@@ -78,7 +77,6 @@ class AIGuardAsync(ServiceBaseAsync):
                 are to be applied to the text, such as defang malicious URLs.
             debug: Setting this value to true will provide a detailed analysis
                 of the text data
-            llm_info: Short string hint for the LLM Provider information
             log_field: Additional fields to include in activity log
 
         Examples:
@@ -92,7 +90,6 @@ class AIGuardAsync(ServiceBaseAsync):
         messages: _T,
         recipe: str | None = None,
         debug: bool | None = None,
-        llm_info: str | None = None,
         log_fields: LogFields | None = None,
     ) -> PangeaResponse[TextGuardResult[_T]]:
         """
@@ -113,48 +110,10 @@ class AIGuardAsync(ServiceBaseAsync):
                 are to be applied to the text, such as defang malicious URLs.
             debug: Setting this value to true will provide a detailed analysis
                 of the text data
-            llm_info: Short string hint for the LLM Provider information
             log_field: Additional fields to include in activity log
 
         Examples:
             response = await ai_guard.guard_text(messages=[{"role": "user", "content": "hello world"}])
-        """
-
-    @overload
-    async def guard_text(
-        self,
-        *,
-        llm_input: _T,
-        recipe: str | None = None,
-        debug: bool | None = None,
-        llm_info: str | None = None,
-        log_fields: LogFields | None = None,
-    ) -> PangeaResponse[TextGuardResult[_T]]:
-        """
-        Text Guard for scanning LLM inputs and outputs
-
-        Analyze and redact text to avoid manipulation of the model, addition of
-        malicious content, and other undesirable data transfers.
-
-        OperationId: ai_guard_post_v1_text_guard
-
-        Args:
-            llm_input: Structured full llm payload data to be scanned by AI
-                Guard for PII, sensitive data, malicious content, and other data
-                types defined by the configuration. Supports processing up to
-                10KB of JSON text
-            recipe: Recipe key of a configuration of data types and settings
-                defined in the Pangea User Console. It specifies the rules that
-                are to be applied to the text, such as defang malicious URLs.
-            debug: Setting this value to true will provide a detailed analysis
-                of the text data
-            llm_info: Short string hint for the LLM Provider information
-            log_field: Additional fields to include in activity log
-
-        Examples:
-            response = await ai_guard.guard_text(
-                llm_input={"model": "gpt-4o", "messages": [{"role": "user", "content": "hello world"}]}
-            )
         """
 
     async def guard_text(  # type: ignore[misc]
@@ -162,10 +121,8 @@ class AIGuardAsync(ServiceBaseAsync):
         text: str | None = None,
         *,
         messages: _T | None = None,
-        llm_input: _T | None = None,
         recipe: str | None = None,
         debug: bool | None = None,
-        llm_info: str | None = None,
         log_fields: LogFields | None = None,
     ) -> PangeaResponse[TextGuardResult[None]]:
         """
@@ -184,27 +141,19 @@ class AIGuardAsync(ServiceBaseAsync):
                 PII, sensitive data, malicious content, and other data types
                 defined by the configuration. Supports processing up to 10KB of
                 JSON text
-            llm_input: Structured full llm payload data to be scanned by AI
-                Guard for PII, sensitive data, malicious content, and other data
-                types defined by the configuration. Supports processing up to
-                10KB of JSON text
             recipe: Recipe key of a configuration of data types and settings
                 defined in the Pangea User Console. It specifies the rules that
                 are to be applied to the text, such as defang malicious URLs.
             debug: Setting this value to true will provide a detailed analysis
                 of the text data
-            llm_info: Short string hint for the LLM Provider information
             log_field: Additional fields to include in activity log
 
         Examples:
             response = await ai_guard.guard_text("text")
         """
 
-        if not any((text, messages, llm_input)):
-            raise ValueError("Exactly one of `text`, `messages`, or `llm_input` must be given")
-
-        if sum((text is not None, messages is not None, llm_input is not None)) > 1:
-            raise ValueError("Only one of `text`, `messages`, or `llm_input` can be given at once")
+        if text is not None and messages is not None:
+            raise ValueError("Exactly one of `text` or `messages` must be given")
 
         return await self.request.post(
             "v1/text/guard",
@@ -212,10 +161,8 @@ class AIGuardAsync(ServiceBaseAsync):
             data={
                 "text": text,
                 "messages": messages,
-                "llm_input": llm_input,
                 "recipe": recipe,
                 "debug": debug,
-                "llm_info": llm_info,
                 "log_fields": log_fields,
             },
         )