+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Redact: `score` in `RecognizerResult` is now a float.
 
+### Removed
+
+- AI Guard: `llm_info` and `llm_input`.
+
 ## 5.5.1 - 2025-02-17
 
 ### Changed